+++ src/libs/core/score/tstaffitem.h
@@ -1,5 +1,5 @@
 /***************************************************************************
- *   Copyright (C) 2017-2018 by Tomasz Bojczuk                             *
+ *   Copyright (C) 2017-2021 by Tomasz Bojczuk                             *
  *   seelook@gmail.com                                                     *
  *                                                                         *
  *   This program is free software; you can redistribute it and/or modify  *
@@ -46,7 +46,7 @@ class NOOTKACORE_EXPORT  TstaffItem : public QQuickItem
   Q_PROPERTY(qreal notesIndent READ notesIndent WRITE setNotesIndent)
   Q_PROPERTY(int firstMeasureNr READ firstMeasureNr NOTIFY firstMeasureNrChanged)
   Q_PROPERTY(int number READ number WRITE setNumber NOTIFY numberChanged)
-  Q_PROPERTY(int scordSpace READ scordSpace WRITE setScordSpace)
+  Q_PROPERTY(qreal scordSpace READ scordSpace WRITE setScordSpace)
 
   friend class TscoreObject;
   friend class TnotePair;
@@ -63,8 +63,8 @@ public:
        * @p 0 - means no scordature (standard tuning)
        * positive value determines additional space below first staff for scordature marks
        */
-  int scordSpace() const { return m_scordSpace; }
-  void setScordSpace(int hasScord);
+  qreal scordSpace() const { return m_scordSpace; }
+  void setScordSpace(qreal hasScord);
 
   TscoreObject* score() { return m_scoreObj; }
   void setScore(TscoreObject* s);
@@ -187,7 +187,7 @@ private:
   qreal                          m_allNotesWidth = 0.0;
   qreal                          m_gapsSum = 0.0;
   QQuickItem                    *m_extraTie = nullptr; /**< Tie at the staff beginning */
-  int                            m_scordSpace = 0;
+  qreal                          m_scordSpace = 0.0;
 };
 
 #endif // TSTAFFITEM_H
